Simple Refactor.
